Refactoring

+++ src/app/patient/show-patient/show-patient.component.ts
@@ -41,10 +41,12 @@ export class ShowPatientComponent implements OnInit {
   deletePatient(del: any){
     del = del.replaceAll('.','').replaceAll('-','');
     try {
+      if(confirm('Você tem certeza que deseja deletar ? ')){
         this.service.deletePatientList(del).subscribe(data => {
           this.refreshPatientList();
           this.showSuccess();
         })
+      }
 
     }catch(err){
       this.showError();
@@ -63,8 +65,11 @@ export class ShowPatientComponent implements OnInit {
     this.service.getPatientList().subscribe(data => {
       this.loading = false;
       this.PatientList = data;
-      for (var i = 0; i < data.length;i++)
+      for (let i = 0; i < data.length;i++){
         data[i].cpf = data[i].cpf.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, "$1.$2.$3-$4")
+        data[i].tel = data[i].tel.replace(/(\d{2})(\d{5})(\d{4})/, "($1) $2-$3")
+        data[i].emergency_tel = data[i].emergency_tel.replace(/(\d{2})(\d{4})(\d{4})/, "($1) $2-$3")
+      }
     })
   }
